+++ app/css/about-us-free-trial.css
@@ -118,3 +118,96 @@
   line-height: normal;
   max-width: 400px;
 }
+@media screen and (max-width: 768px) {
+  .about-us-free-trial {
+    min-height: 100vh; /* Ensures the section takes full viewport height */
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding-bottom: 0;
+    position: relative;
+  }
+
+  .free-trial-container {
+    width: 100%;
+    height: 100%;
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .free-trial-image {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: 1;
+  }
+
+  .free-trial-content-block {
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 2rem;
+    position: relative;
+    z-index: 2;
+    width: 100%;
+    padding: 2rem;
+  }
+
+  .free-trial-content {
+    text-align: center;
+    width: 100%;
+  }
+
+  .free-trial-title {
+    font-size: 1.8rem;
+    max-width: 100%;
+  }
+
+  .free-trial-description {
+    font-size: 1rem;
+    width: 100%;
+  }
+
+  .free-trial-time {
+    width: 100%;
+    padding: 10px;
+    font-size: 1rem;
+  }
+
+  .free-trial-drop-down {
+    width: 100%;
+    font-size: 1rem;
+    height: 50px;
+  }
+
+  .trial-offer {
+    width: 100%;
+    height: auto;
+    padding: 2rem;
+    text-align: center;
+    background: rgba(99, 174, 166, 0.7);
+    backdrop-filter: blur(10px);
+    border-radius: 15px;
+  }
+
+  .trial-offer-features-block {
+    gap: 1rem;
+  }
+
+  .trial-offer-features-line {
+    justify-content: start;
+    gap: 10px;
+    text-align: start;
+  }
+
+  .trial-offer-features-paragraph {
+    font-size: 1rem;
+  }
+}
