+++ app/css/about-us-guidance.css
@@ -31,3 +31,46 @@
 .guidance-image{
   z-index: 9;
 }
+
+
+@media (max-width: 768px) {
+  .about-us-guidance-section {
+    padding: 2rem;
+  }
+
+  .about-us-guidance-container {
+    flex-direction: column;
+    text-align: center;
+    gap: 3rem;
+  }
+
+  .guidance-content {
+    width: 100%;
+    align-items: center;
+    gap: 1.5rem;
+  }
+
+  .section-title {
+    font-size: 2rem;
+    line-height: 1.3;
+  }
+
+  .guidance-description {
+    width: 100%;
+    font-size: 1.25rem;
+    line-height: 3vh;
+    text-align: justify;
+  }
+
+  .guidance-image {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .guidance-image img {
+    width: 100%;
+    max-width: 350px;
+    height: auto;
+  }
+}
