+++ app/css/howwework.css
@@ -38,3 +38,22 @@
     display: flex;
     gap: 28px;
 }
+
+@media screen and (max-width: 768px) {
+    .cards-outer-div {
+      flex-direction: column; 
+      align-items: center; 
+      gap: 20px;
+    }
+  
+    .card-content {
+      flex-direction: column; 
+      align-items: center; 
+      gap: 20px;
+    }
+  
+    .top-arrow, .bottom-arrow {
+      display: none; 
+    }
+  }
+
