+++ app/css/our-course-teaching.css
@@ -4,6 +4,24 @@
   padding: 45px auto;
   background-color: var(--background);
 }
+.our-courses-page-grid-image-left {
+  position: absolute;
+  top: 10px;
+  left: 0;
+  z-index: 3;
+  width: 25.25rem;
+  height: 93.50013rem;
+  flex-shrink: 0;
+}
+.our-courses-page-grid-image-right {
+  position: absolute;
+  top: 10px;
+  right: 0;
+  z-index: 3;
+  width: 25.25rem;
+  height: 93.50013rem;
+  flex-shrink: 0;
+}
 .teaching-content {
   position: relative;
   display: flex;
@@ -13,8 +31,8 @@
   justify-content: space-between;
   align-items: center;
   gap: 10px;
-  /* gap: 69px; */
   border-radius: 20px;
+  z-index: 15;
   background: linear-gradient(4deg, #204343 2.5%, #42736b 79.71%);
 }
 .teaching-content-title {
@@ -34,4 +52,32 @@ font-style: normal;
 font-weight: 600;
 line-height: normal;
 max-width: 587px;
+}
+
+@media screen and (max-width: 768px) {
+  .our-courses-page-grid-image-left{
+    display: none;
+  }
+  .our-courses-page-grid-image-right{
+    display: none;
+  }
+  .teaching-content {
+    flex-direction: column;
+    width: 90%; 
+    height: auto;
+    padding: 20px;
+    text-align: center; 
+    gap: 30px;
+    margin: 0 auto; 
+  }
+
+  .teaching-content-title {
+    font-size: 26px; 
+    max-width: 100%; 
+  }
+
+  .teaching-content-description {
+    font-size: 18px;
+    max-width: 100%;
+  }
 }
