+++ app/css/our-courses-live-card.css
@@ -1,0 +1,143 @@
+.our-courses-card {
+  width: 34.375rem;
+  height: 50rem;
+  flex-shrink: 0;
+  border-radius: 1.25rem;
+  background: #fff;
+}
+.courses-image {
+  width: 34.375rem;
+  height: 31.25rem;
+  border-radius: 1.25rem;
+  background: linear-gradient(
+    0deg,
+    rgba(29, 29, 29, 0.2) 0%,
+    rgba(29, 29, 29, 0.2) 100%
+  );
+}
+.our-courses-card-image {
+  width: 34.375rem;
+height: 31.25rem;
+  object-fit: cover;
+  border-radius: 1.25rem;
+}
+.our-courses-content {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  align-items: start;
+  justify-content: start;
+  margin: 10px 20px;
+}
+.our-courses-subtitle {
+  color: #000;
+  font-family: Inter;
+  font-size: 1.25rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+}
+.our-courses-title {
+  color: #204343;
+  font-family: Inter;
+  font-size: 2.5rem;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
+.our-courses-description-text {
+  color: #000;
+  font-family: Inter;
+  font-size: 1.125rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  max-width: 18.0625rem;
+}
+.courses-btn {
+  display: flex;
+  width: 17.5rem;
+  height: 3.125rem;
+  padding: 0.3125rem 0.6875rem 0.3125rem 1.3125rem;
+  align-items: center;
+  gap: 0.5rem;
+  flex-shrink: 0;
+  border-radius: 0.625rem;
+  background: #204343;
+}
+.courses-btn-text {
+  color: #fff;
+  font-family: Inter;
+  font-size: 1.25rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  width: 13.0625rem;
+  flex-shrink: 0;
+}
+.courses-btn-icon {
+  width: 1.25rem;
+  height: 0rem;
+  flex-shrink: 0;
+  stroke-width: 2px;
+stroke: #FFF;
+}
+
+@media screen and (max-width: 768px) {
+  .our-courses-card {
+    width: 90%;
+    height: auto;
+    border-radius: 1rem;
+  }
+
+  .courses-image {
+    width: 100%;
+    height: auto;
+    border-radius: 1rem;
+  }
+
+  .our-courses-card-image {
+    width: 100%;
+    height: auto;
+    object-fit: cover;
+    border-radius: 1rem;
+  }
+
+  .our-courses-content {
+    gap: 10px;
+    margin: 15px;
+    text-align: center;
+    align-items: center;
+  }
+
+  .our-courses-subtitle {
+    font-size: 1rem;
+  }
+
+  .our-courses-title {
+    font-size: 2rem;
+  }
+
+  .our-courses-description-text {
+    font-size: 1rem;
+    max-width: 100%;
+  }
+
+  .courses-btn {
+    width: 100%;
+    height: 3rem;
+    padding: 0.5rem 1rem;
+    gap: 0.5rem;
+    border-radius: 0.5rem;
+  }
+
+  .courses-btn-text {
+    font-size: 1rem;
+    width: auto;
+  }
+
+  .courses-btn-icon {
+    width: 1.5rem;
+    height: auto;
+  }
+}
