+++ app/css/our-journey.css
@@ -1,0 +1,191 @@
+.our-journey-section {
+  padding: 45px 0 0 0;
+  z-index: 5;
+  background-color: var(--background);
+}
+.our-journey-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 95px;
+}
+.our-journey-content {
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  justify-content: center;
+  gap: 30px;
+}
+.our-journey-title {
+  max-width: 430px;
+  color: #1D1D1D;
+font-family: "DM Sans";
+font-size: 45px;
+font-style: normal;
+font-weight: 600;
+line-height: normal;
+}
+.our-journey-title span {
+    color: #63AEA7;
+font-family: "DM Sans";
+font-size: 45px;
+font-style: normal;
+font-weight: 600;
+line-height: normal;
+}
+.our-journey-description {
+  max-width: 549px;
+  color: #1d1d1d;
+  font-family: "DM Sans";
+  font-size: 35px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+}
+
+.stats-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #f5f5f5;
+  /* padding: 20px; */
+}
+
+.stats-wrapper {
+  position: relative;
+  width: 600px;
+  height: 600px;
+}
+
+.circle {
+  position: absolute;
+  width: 150px;
+  height: 150px;
+  border-radius: 50%;
+  background-color: #5fbfb1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  color: white;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  transition: transform 0.3s ease;
+}
+
+.circle:hover {
+  transform: scale(1.05);
+}
+
+.circle.center {
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: #1d4942;
+  z-index: 2;
+}
+
+.circle.center:hover {
+  transform: translate(-50%, -50%) scale(1.05);
+}
+
+.circle.top-left {
+  top: 10%;
+  left: 20%;
+}
+
+.circle.top-right {
+  top: 10%;
+  right: 20%;
+}
+
+.circle.middle-left {
+  top: 50%;
+  left: 5%;
+  transform: translateY(-50%);
+}
+
+.circle.middle-left:hover {
+  transform: translateY(-50%) scale(1.05);
+}
+
+.circle.middle-right {
+  top: 50%;
+  right: 5%;
+  transform: translateY(-50%);
+}
+
+.circle.middle-right:hover {
+  transform: translateY(-50%) scale(1.05);
+}
+
+.circle.bottom-left {
+  bottom: 10%;
+  left: 20%;
+}
+
+.circle.bottom-right {
+  bottom: 10%;
+  right: 20%;
+}
+
+.content {
+  padding: 15px;
+}
+
+.number {
+  font-size: 24px;
+  font-weight: bold;
+  margin-bottom: 5px;
+}
+
+.label {
+  font-size: 16px;
+}
+
+/* Responsive adjustments */
+@media screen and (max-width: 768px) {
+  .our-journey-container {
+    flex-direction: column;
+    gap: 50px;
+    text-align: center;
+  }
+
+  .our-journey-content {
+    align-items: center;
+  }
+
+  .our-journey-title,
+  .our-journey-title span {
+    font-size: 32px;
+  }
+
+  .our-journey-description {
+    font-size: 20px;
+    max-width: 100%;
+    padding: 0 20px;
+  }
+
+  .our-journey-cirlce-block{
+    /* display: none; */
+    /* height: 500px;
+    width: 400px; */
+  }
+  .stats-wrapper{
+    height: 450px;
+    width: 420px;
+  }
+  .circle {
+    position: absolute;
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    background-color: #5fbfb1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    color: white;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease;
+  }
+}
